@@ -1,10 +1,10 @@
 package com.wa2c.android.medoly.plugin.action.lrclyrics.dialog
 
-import android.app.AlertDialog
 import android.app.Dialog
-import android.databinding.DataBindingUtil
 import android.os.Bundle
 import android.view.LayoutInflater
+import androidx.appcompat.app.AlertDialog
+import androidx.databinding.DataBindingUtil
 import com.wa2c.android.medoly.plugin.action.lrclyrics.R
 import com.wa2c.android.medoly.plugin.action.lrclyrics.databinding.DialogCacheBinding
 import com.wa2c.android.medoly.plugin.action.lrclyrics.db.SearchCache
@@ -26,10 +26,10 @@ class CacheDialogFragment : AbstractDialogFragment() {
 
     override fun onCreateDialog(savedInstanceState: Bundle?): Dialog {
         super.onCreateDialog(savedInstanceState)
-        binding = DataBindingUtil.inflate(LayoutInflater.from(activity), R.layout.dialog_cache, null, false)
+        binding = DataBindingUtil.inflate(LayoutInflater.from(context), R.layout.dialog_cache, null, false)
 
         // data
-        val cache = arguments.getSerializable(ARG_CACHE) as SearchCache
+        val cache = arguments!!.getSerializable(ARG_CACHE) as SearchCache
         val result = cache.makeResultItem()
 
         // view
@@ -47,7 +47,7 @@ class CacheDialogFragment : AbstractDialogFragment() {
         }
 
         // build dialog
-        val builder = AlertDialog.Builder(activity)
+        val builder = AlertDialog.Builder(context)
         builder.setTitle(R.string.title_activity_cache)
         builder.setView(binding.root)
         builder.setNeutralButton(R.string.label_close, null)
@@ -62,7 +62,7 @@ class CacheDialogFragment : AbstractDialogFragment() {
      * Delete cache lyrics.
      */
     private fun deleteLyrics(cache: SearchCache) {
-        val searchCacheHelper = SearchCacheHelper(this@CacheDialogFragment.activity)
+        val searchCacheHelper = SearchCacheHelper(context)
         GlobalScope.launch(Dispatchers.Main) {
             val deleteResult = async(Dispatchers.Default) {
                 try {
@@ -73,7 +73,7 @@ class CacheDialogFragment : AbstractDialogFragment() {
             }
             val r = deleteResult.await()
             if (r != true)
-                AppUtils.showToast(this@CacheDialogFragment.activity, R.string.message_dialog_cache_delete_error)
+                AppUtils.showToast(context, R.string.message_dialog_cache_delete_error)
             clickListener?.invoke(dialog, DIALOG_RESULT_DELETE_LYRICS, null)
             dialog.dismiss()
         }
@@ -83,7 +83,7 @@ class CacheDialogFragment : AbstractDialogFragment() {
      * Delete cache.
      */
     private fun deleteCache(cache: SearchCache) {
-        val searchCacheHelper = SearchCacheHelper(this@CacheDialogFragment.activity)
+        val searchCacheHelper = SearchCacheHelper(context)
         GlobalScope.launch(Dispatchers.Main) {
             val deleteResult = async(Dispatchers.Default) {
                 try {
@@ -94,7 +94,7 @@ class CacheDialogFragment : AbstractDialogFragment() {
             }
             val r = deleteResult.await()
             if (r != true)
-                AppUtils.showToast(this@CacheDialogFragment.activity, R.string.message_dialog_cache_delete_error)
+                AppUtils.showToast(context, R.string.message_dialog_cache_delete_error)
             clickListener?.invoke(dialog, DIALOG_RESULT_DELETE_CACHE, null)
             dialog.dismiss()
         }
